refactor(admin): move weaver edits editor into a metadata-style route drawer
Replaces the inline WeaverEditsSection with the idiomatic RouteDrawer pattern.

- New @edits parallel route renders the editor in a slide-over drawer.
- WeaverEditsForm wraps the social media / corrections / custom-fields editor
  in RouteDrawer, closing via useRouteModal().handleSuccess() on save.
- The weaver detail page now renders an outlet and a small "Edits & changes"
  summary with an Edit action that opens the drawer.
