+++ apps/backend/src/admin/routes/persons/weavers/[census_id]/@edits/page.tsx
@@ -1,0 +1,11 @@
+import { useParams } from "react-router-dom"
+
+import { WeaverEditsForm } from "../../../../../components/persons/weaver-edits-form"
+
+const WeaverEditsDrawer = () => {
+  const { census_id } = useParams()
+
+  return <WeaverEditsForm censusId={census_id!} />
+}
+
+export default WeaverEditsDrawer
